requirements.txt
flask
pymongo
python-dotenv
werkzeug
google-genai
